We now succesfully read a gltf scene

+++ src/engine/Mesh.hpp
@@ -4,5 +4,5 @@ public:
 	std::vector<glm::vec3>  vertexPositions;
 	std::vector<glm::vec3>  normals;
 	std::vector<glm::vec2>  uvCoords;
-	std::vector<int>		faceIndices;
+	std::vector<short>		faceIndices;
 };
